consume useFidesGlobal in overlay rendering
