+++ clients/fides-js/src/components/types.ts
@@ -1,16 +1,8 @@
 import { ContainerNode } from "preact";
 
-import type {
-  FidesConfig,
-  FidesCookie,
-  FidesExperienceTranslationOverrides,
-  FidesGlobal,
-  FidesInitOptions,
-  NoticeConsent,
-  PrivacyExperience,
-  PrivacyExperienceMinimal,
-} from "../lib/consent-types";
+import { FidesExperienceTranslationOverrides } from "../lib/consent-types";
 import type { I18n } from "../lib/i18n";
+import { InitializedFidesGlobal } from "../lib/providers/fides-global-context";
 
 /**
  * The type of the parent component for the preact app
@@ -19,26 +11,11 @@ import type { I18n } from "../lib/i18n";
  * Similarly, when creating different overlay "types", they should all take
  * this type as a prop.
  */
-export interface OverlayProps {
-  options: FidesInitOptions;
-  experience: PrivacyExperience | PrivacyExperienceMinimal;
-  cookie: FidesCookie;
-  fidesRegionString: string;
-  savedConsent: NoticeConsent;
-  translationOverrides?: Partial<FidesExperienceTranslationOverrides>;
-}
-
 export type RenderOverlayType = (
-  props: OverlayProps & {
+  props: {
     i18n: I18n;
     initializedFides: InitializedFidesGlobal;
+    translationOverrides?: Partial<FidesExperienceTranslationOverrides>;
   },
   parent: ContainerNode,
 ) => void;
-
-export interface InitializedFidesGlobal extends FidesGlobal {
-  cookie: FidesCookie;
-  config: FidesConfig;
-  experience: PrivacyExperience | PrivacyExperienceMinimal;
-  fidesRegionString: string;
-}
