+++ clients/fides-js/src/lib/renderOverlay.tsx
@@ -1,20 +1,18 @@
 import { h, render } from "preact";
 
-import NoticeOverlay, {
-  NoticeOverlayProps,
-} from "../components/notices/NoticeOverlay";
+import NoticeOverlay from "../components/notices/NoticeOverlay";
 import { RenderOverlayType } from "../components/types";
 import { I18nProvider } from "./i18n/i18n-context";
 import { EventProvider } from "./providers/event-context";
 import { FidesGlobalProvider } from "./providers/fides-global-context";
 
 export const renderOverlay: RenderOverlayType = (props, parent) => {
-  const { i18n, initializedFides, ...overlayProps } = props;
+  const { i18n, initializedFides } = props;
   render(
     <I18nProvider i18nInstance={i18n}>
       <FidesGlobalProvider initializedFides={initializedFides}>
         <EventProvider>
-          <NoticeOverlay {...(overlayProps as NoticeOverlayProps)} />
+          <NoticeOverlay />
         </EventProvider>
       </FidesGlobalProvider>
     </I18nProvider>,
